% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/180709_sex2dex.R
\name{ch2wgs}
\alias{ch2wgs}
\title{Convert CH coordinates to WGS}
\usage{
ch2wgs(x, y)
}
\arguments{
\item{x}{the value of the xaxis}

\item{y}{the value of the yaxis}
}
\description{
This function allows you to convert the swiss based coordinates to latitute and longitute
}
\examples{
ch2wgs()
}
