% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/180709_dex2sex.R
\name{dec2sex}
\alias{dec2sex}
\title{Convert decimal degrees to sexagesimal seconds}
\usage{
dec2sex(angle)
}
\arguments{
\item{angle}{An vector of angles to be converted to sexagesimal second}
}
\description{
This function allows you convert decimal degrees to sexagesimal seconds
}
\examples{
cat_function()
}
