% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/180711_DataDescription_1.R
\docType{data}
\name{LandShapefile}
\alias{LandShapefile}
\title{Shapefiles of Switzerland in 2017}
\format{An object of class SpatialPolygonDataFrame}
\source{
https://www.swisstopo.admin.ch/
}
\usage{
data(LandShapefile)
data(CantonShapefile)
data(MunicipalityShapefile)
}
\description{
Data was retrieved by https://www.swisstopo.admin.ch/ and includes 
the shapefiles of the land, cantons and municipalities of Switzerland in 2017.
}
\examples{
data(CantonShapefile)
plot(CantonShapefile)
# to get an idea about the crude population counts
spplot(pol_canton2017, "EINWOHNERZ")
# and population density
pol_canton2017$popdens <- pol_canton2017$EINWOHNERZ/area(pol_canton2017)
spplot(pol_canton2017, "popdens")
}
\keyword{datasets}
