% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/180709_RemoveUmlaut.R
\name{rmuml}
\alias{rmuml}
\title{An Umlaut function}
\usage{
rmuml(x)
}
\arguments{
\item{x}{a character type word with umlauts}
}
\description{
This function identifies the umlaut of a word and replaces it with the same word omitting the umlauts
}
\examples{
rmuml()
}
