% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/180709_Long2Swiss.R
\name{WGS2CH}
\alias{WGS2CH}
\title{A WGS lat/long to CH}
\usage{
WGS2CH(lat, lng)
}
\arguments{
\item{lat}{A vector of latitute values}

\item{long}{A vector of longitute values}
}
\description{
This function allows you to covert the latitude and longitute to swiss coordinates
}
\examples{
WGS2CH()
}
